[editor] Added picking (needs perspective correction), other changes
Renamed ImGuiToastType to ENotifType
Entity deletion crash fixed
Scene name displayed,
Added fbx drag and drop

+++ CherryCrisis/CherryEngine/include/entity.hpp
@@ -62,7 +62,7 @@ public:
 
 	void SetName(std::string name) { m_name = name; }
 	std::string GetName() { return m_name; }
-	uint64_t	GetUUID() { return (uint64_t)m_uuid; }
+	uint32_t	GetUUID() { return (uint32_t)m_uuid; }
 	std::string Serialized();
 	std::string SerializeBehaviours();
 };
